masukkan model de file aplikasi

+++ app/src/main/java/com/example/healthcare/domain/model/DiagnosisHistory.kt
@@ -3,7 +3,7 @@ package com.example.healthcare.domain.model
 data class DiagnosisHistory(
     val id: String,
     val penyakit: String,
-    val persentase: Int,
+    val persentase: Double,
     val tanggal: String,
     val waktu: String
 )
